+++ src/Helper/Order.php
@@ -24,6 +24,12 @@ namespace Shopgate\Import\Helper;
 
 use Magento\Framework\Api\SimpleDataObjectConverter;
 use Magento\Framework\Event\ManagerInterface;
+use Magento\Framework\Exception\CouldNotSaveException;
+use Magento\Framework\Exception\InputException;
+use Magento\Framework\Exception\LocalizedException;
+use Magento\Framework\Exception\MailException;
+use Magento\Framework\Exception\NoSuchEntityException;
+use Magento\Framework\Serialize\SerializerInterface;
 use Magento\Quote\Api\CartManagementInterface;
 use Magento\Quote\Model\QuoteManagement;
 use Magento\Quote\Model\QuoteRepository;
@@ -38,6 +44,7 @@ use Shopgate\Base\Model\Utility\SgLoggerInterface;
 use Shopgate\Import\Helper\Order\Shipping;
 use Shopgate\Import\Model\Payment\Factory as PaymentFactory;
 use Shopgate\Import\Model\Service\Import as ImportService;
+use ShopgateLibraryException;
 
 class Order
 {
@@ -71,6 +78,8 @@ class Order
     private $shippingHelper;
     /** @var PaymentFactory */
     private $paymentFactory;
+    /** @var SerializerInterface */
+    private $serializer;
 
     /**
      * @param Base                     $order
@@ -87,6 +96,7 @@ class Order
      * @param Shipping                 $shippingHelper
      * @param ManagerInterface         $eventManager
      * @param PaymentFactory           $paymentFactory
+     * @param SerializerInterface      $serializer
      * @param array                    $quoteMethods
      */
     public function __construct(
@@ -104,6 +114,7 @@ class Order
         Shipping $shippingHelper,
         ManagerInterface $eventManager,
         PaymentFactory $paymentFactory,
+        SerializerInterface $serializer,
         array $quoteMethods = []
     ) {
         $this->sgOrder           = $order;
@@ -121,6 +132,7 @@ class Order
         $this->shippingHelper    = $shippingHelper;
         $this->eventManager      = $eventManager;
         $this->paymentFactory    = $paymentFactory;
+        $this->serializer        = $serializer;
     }
 
     /**
@@ -155,11 +167,11 @@ class Order
     /**
      * Creates the order then we can continue loading on $this->mageOrder
      *
-     * @throws \Magento\Framework\Exception\CouldNotSaveException
-     * @throws \Magento\Framework\Exception\InputException
-     * @throws \Magento\Framework\Exception\NoSuchEntityException
-     * @throws \ShopgateLibraryException
-     * @throws \Magento\Framework\Exception\LocalizedException
+     * @throws CouldNotSaveException
+     * @throws InputException
+     * @throws NoSuchEntityException
+     * @throws ShopgateLibraryException
+     * @throws LocalizedException
      */
     protected function setStartAdd()
     {
@@ -175,7 +187,7 @@ class Order
         $this->eventManager->dispatch('checkout_submit_before', ['quote' => $quote]);
         $order = $this->quoteManagement->submit($quote);
         if (null === $order) {
-            throw new \ShopgateLibraryException(\ShopgateLibraryException::PLUGIN_ORDER_ITEM_NOT_FOUND);
+            throw new ShopgateLibraryException(ShopgateLibraryException::PLUGIN_ORDER_ITEM_NOT_FOUND);
         }
         $this->mageOrder = $order;
         $this->eventManager->dispatch('checkout_submit_all_after', ['order' => $this->mageOrder, 'quote' => $quote]);
@@ -192,9 +204,9 @@ class Order
     /**
      * Updates the order
      *
-     * @throws \Magento\Framework\Exception\NoSuchEntityException
-     * @throws \ShopgateLibraryException
-     * @throws \Magento\Framework\Exception\InputException
+     * @throws NoSuchEntityException
+     * @throws ShopgateLibraryException
+     * @throws InputException
      */
     protected function setStartUpdate()
     {
@@ -202,7 +214,7 @@ class Order
         $this->log->debug('## Order-Number: ' . $orderNumber);
         $this->localSgOrder = $this->sgOrderRepository->checkOrderExists($orderNumber);
         if (!$this->localSgOrder->getShopgateOrderId()) {
-            throw new \ShopgateLibraryException(\ShopgateLibraryException::PLUGIN_ORDER_NOT_FOUND);
+            throw new ShopgateLibraryException(ShopgateLibraryException::PLUGIN_ORDER_NOT_FOUND);
         }
 
         $this->mageOrder = $this->orderRepository->get($this->localSgOrder->getOrderId());
@@ -224,14 +236,18 @@ class Order
      */
     protected function setOrderPayment()
     {
-        $this->paymentFactory->getPayment($this->sgOrder->getPaymentMethod())
-                             ->manipulateOrderWithPaymentData($this->mageOrder, $this->sgOrder);
+        $payment = $this->paymentFactory->getPayment($this->sgOrder->getPaymentMethod());
+        $payment->manipulateOrderWithPaymentDataBeforeSave($this->mageOrder, $this->sgOrder);
+
+        $this->mageOrder = $this->orderRepository->save($this->mageOrder);
+
+        $payment->manipulateOrderWithPaymentDataAfterSave($this->mageOrder, $this->sgOrder);
     }
 
     /**
      * Checks if shipments should be updated for an existing order
      *
-     * @throws \Magento\Framework\Exception\LocalizedException
+     * @throws LocalizedException
      */
     protected function setUpdateShipping()
     {
@@ -248,7 +264,7 @@ class Order
     /**
      * Set correct order status by payment
      *
-     * @throws \Magento\Framework\Exception\LocalizedException
+     * @throws LocalizedException
      */
     protected function setOrderState()
     {
@@ -278,7 +294,7 @@ class Order
     /**
      * Send order notification if activated in config
      *
-     * @throws \Magento\Framework\Exception\MailException
+     * @throws MailException
      */
     protected function setOrderNotification()
     {