+++ src/Model/Payment/Shopgate/Shopgate.php
@@ -30,9 +30,9 @@ use Shopgate\Import\Model\Payment\AbstractPayment;
 
 class Shopgate extends AbstractPayment
 {
-    const MODULE_NAME            = 'Shopgate_Base';
-    const PAYMENT_CODE           = 'shopgate';
-    const XML_CONFIG_STATUS_PAID = 'payment/shopgate/order_status';
+    const MODULE_NAME             = 'Shopgate_Base';
+    const PAYMENT_CODE            = 'shopgate';
+    const XML_CONFIG_ORDER_STATUS = 'payment/shopgate/order_status';
 
     /**
      * Always valid as it is the fallback method
@@ -47,8 +47,10 @@ class Shopgate extends AbstractPayment
     /**
      * @inheritDoc
      */
-    public function manipulateOrderWithPaymentData(MagentoOrder $magentoOrder, ShopgateOrder $shopgateOrder): void
-    {
+    public function manipulateOrderWithPaymentDataBeforeSave(
+        MagentoOrder $magentoOrder,
+        ShopgateOrder $shopgateOrder
+    ): void {
         if ($shopgateOrder->getIsPaid()
             && $magentoOrder->getBaseTotalDue()
             && $magentoOrder->getPayment()
@@ -56,7 +58,7 @@ class Shopgate extends AbstractPayment
             $magentoOrder->getPayment()->setShouldCloseParentTransaction(true);
             $magentoOrder->getPayment()->registerCaptureNotification($shopgateOrder->getAmountComplete());
             $magentoOrder->addStatusHistoryComment(__('[SHOPGATE] Payment received.'))
-                               ->setIsCustomerNotified(false);
+                         ->setIsCustomerNotified(false);
         }
     }
 }